+++ CHANGELOG.md
@@ -1,9 +1,20 @@
+## v0.1.2
+
+- improve emoji command using embed
+- add some stuff to about command
+- improve help command using embed and help for specified command
+
 ## v0.1.1
 
-- fix emote command unknown emote ordered that one line length.
+- fix emote command unknown emote ordered that one line length
 - add `about` command
 - change success, warning and error emoji in response
+- add `pause` command
+- add `skipto` command
+- change the nowplaying embed thumbnail to better quality
+- adding thumbnail and footer source to start playing embed
+- adding duration to start playing embed
 
 ## v0.1.0
 
-- first release information can be found [here](https://github.com/SharifPoetra/thunder-java/releases/tag/0.1.0).
+- first release information can be found [on the releases](https://github.com/SharifPoetra/thunder-java/releases/tag/0.1.0).